refac as prep for sentinel support in AppConfig
